requirements.txt
# Flask核心依赖
Flask==3.0.0

# 数据库相关
Flask-SQLAlchemy==3.1.1
pymysql==1.1.0

# JWT认证
Flask-JWT-Extended==4.6.0

# CORS支持
Flask-CORS==4.0.0

# 数据处理
pandas==2.1.4
openpyxl==3.1.2

# 环境变量管理
python-dotenv==1.0.0

# WebSocket支持
Flask-SocketIO==5.3.6

# 其他
uv==0.1.21